+++ CHANGELOG.md
@@ -9,6 +9,7 @@ General fixes:
 
 get-command:
 * Fix: List options (e.g. `--feature`) will now be shown as `--feature f1 --feature f2` instead of as `--feature ['f1', 'f2']`.
+* Fix: `date` and `time` is no longer shown as the submit command automatically adds these.
 
 ## v1.4.0